+++ src/market_sentinel_ai/features/__init__.py
@@ -1,4 +1,5 @@
+from market_sentinel_ai.features.multitimeframe import aggregate_candles
 from market_sentinel_ai.features.ohlcv import OHLCVFeatureEngine
+from market_sentinel_ai.features.order_flow import OrderFlowFeatureEngine
 
-__all__ = ["OHLCVFeatureEngine"]
-
+__all__ = ["OHLCVFeatureEngine", "OrderFlowFeatureEngine", "aggregate_candles"]